feat(nfcoreUtils): Improve linking and unlinking functionality
