+++ tests/nfcoreInitialise/main.nf
